# 1.48x Faster Partial Caching - Alternative to #36310