--- a/actionview/test/template/render_test.rb
+++ b/actionview/test/template/render_test.rb
@@ -15,7 +15,7 @@ module RenderTestCases
     @view = Class.new(ActionView::Base.with_empty_template_cache) do
       def view_cache_dependencies; []; end
 
-      def combined_fragment_cache_key(key = nil)
+      def combined_fragment_cache_key(key)
         [ :views, key ].compact
       end
     end.with_view_paths(paths, @assigns)