--- a/activesupport/lib/active_support/cache.rb
+++ b/activesupport/lib/active_support/cache.rb
@@ -9,7 +9,6 @@ require "active_support/core_ext/numeric/time"
 require "active_support/core_ext/object/to_param"
 require "active_support/core_ext/object/try"
 require "active_support/core_ext/string/inflections"
-require "active_support/cache/key"
 
 module ActiveSupport
   # See ActiveSupport::Cache::Store for documentation.
@@ -87,11 +86,21 @@ module ActiveSupport
           expanded_cache_key << "#{prefix}/"
         end
 
-        expanded_cache_key << ActiveSupport::Cache::Key.cache_key_with_version(key)
+        expanded_cache_key << retrieve_cache_key(key)
         expanded_cache_key
       end
 
       private
+        def retrieve_cache_key(key)
+          case
+          when key.respond_to?(:cache_key_with_version) then key.cache_key_with_version
+          when key.respond_to?(:cache_key)              then key.cache_key
+          when key.is_a?(Array)                         then key.map { |element| retrieve_cache_key(element) }.to_param
+          when key.respond_to?(:to_a)                   then retrieve_cache_key(key.to_a)
+          else                                               key.to_param
+          end.to_s
+        end
+
         # Obtains the specified cache store class, given the name of the +store+.
         # Raises an error when the store class cannot be found.
         def retrieve_store_class(store)
@@ -306,7 +315,6 @@ module ActiveSupport
       #   cache.fetch('foo') # => "bar"
       def fetch(name, options = nil)
         if block_given?
-          name = ActiveSupport::Cache::Key(name)
           options = merged_options(options)
           key = normalize_key(name, options)
 
@@ -342,7 +350,6 @@ module ActiveSupport
       # Options are passed to the underlying cache implementation.
       def read(name, options = nil)
         options = merged_options(options)
-        name    = ActiveSupport::Cache::Key(name)
         key     = normalize_key(name, options)
         version = normalize_version(name, options)
 
@@ -453,7 +460,7 @@ module ActiveSupport
       # Options are passed to the underlying cache implementation.
       def write(name, value, options = nil)
         options = merged_options(options)
-        name = ActiveSupport::Cache::Key(name)
+
         instrument(:write, name, options) do
           entry = Entry.new(value, **options.merge(version: normalize_version(name, options)))
           write_entry(normalize_key(name, options), entry, **options)
@@ -664,8 +671,24 @@ module ActiveSupport
           end
         end
 
+        # Expands key to be a consistent string value. Invokes +cache_key+ if
+        # object responds to +cache_key+. Otherwise, +to_param+ method will be
+        # called. If the key is a Hash, then keys will be sorted alphabetically.
         def expanded_key(key)
-          ActiveSupport::Cache::Key(key).cache_key
+          return key.cache_key.to_s if key.respond_to?(:cache_key)
+
+          case key
+          when Array
+            if key.size > 1
+              key = key.collect { |element| expanded_key(element) }
+            else
+              key = expanded_key(key.first)
+            end
+          when Hash
+            key = key.sort_by { |k, _| k.to_s }.collect { |k, v| "#{k}=#{v}" }
+          end
+
+          key.to_param
         end
 
         def normalize_version(key, options = nil)
@@ -673,7 +696,11 @@ module ActiveSupport
         end
 
         def expanded_version(key)
-          ActiveSupport::Cache::Key(key).cache_version
+          case
+          when key.respond_to?(:cache_version) then key.cache_version.to_param
+          when key.is_a?(Array)                then key.map { |element| expanded_version(element) }.compact.to_param
+          when key.respond_to?(:to_a)          then expanded_version(key.to_a)
+          end
         end
 
         def instrument(operation, key, options = nil)
@@ -824,3 +851,4 @@ module ActiveSupport
     end
   end
 end
+require "active_support/cache/key"