--- a/activesupport/lib/active_support/cache/key.rb
+++ b/activesupport/lib/active_support/cache/key.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'active_support/cache'
+require "active_support/cache"
 
 module ActiveSupport
   module Cache