--- a/activesupport/lib/active_support/cache/key.rb
+++ b/activesupport/lib/active_support/cache/key.rb
@@ -1,27 +1,9 @@
 # frozen_string_literal: true
 
+require 'active_support/cache'
+
 module ActiveSupport
   module Cache
-    # Convenience method to prevent re-generating a Key
-    # instance, if the element being passed in is already
-    # instance of the `Key` class.
-    #
-    # Example:
-    #
-    #   ActiveSupport::Cache::Key("foo").cache_key
-    #    # => "foo"
-    #
-    #   k1 = Key.new("foo")
-    #   key = ActiveSupport::Cache::Key(k1)
-    #   k1.object_id == key.object_id # => true
-    def self.Key(key_or_components) # :nodoc:
-      if key_or_components.is_a?(ActiveSupport::Cache::Key)
-        key_or_components
-      else
-        Key.new(key_or_components)
-      end
-    end
-
     # This class is responsible for coercing input into a valid
     # key format and cache version for cache stores.
     #
@@ -111,92 +93,44 @@ module ActiveSupport
     #   Key.new({foo: klass.new}).cache_version
     #   # => "foo"
     class Key # :nodoc:
-      attr_reader :cache_key, :cache_version
+      DEFAULT_KEY = Object.new
+      CACHE_METHOD_OBJ = ActiveSupport::Cache::Store.new
 
-      def initialize(key = nil)
-        @cache_key = +""
-        @cache_version = +""
-        @empty_key = true
+
+      def initialize(key = DEFAULT_KEY)
+        @cache_key = nil
+        @cache_version = nil
         @key_parts = []
-        self << key
+        self << key unless key == DEFAULT_KEY
+      end
+
+      def cache_key
+        @cache_key ||= begin
+          CACHE_METHOD_OBJ.send(:expanded_key, @key_parts)
+        end
+      end
+
+      def cache_version
+        @cache_version ||= begin
+          CACHE_METHOD_OBJ.send(:expanded_version, @key_parts)
+        end
       end
 
       def length
-        @cache_key.length
+        cache_key.length
       end
 
-      def self.cache_key_with_version(key)
-        case
-        when key.respond_to?(:cache_key_with_version) then key.cache_key_with_version
-        when key.respond_to?(:cache_key)              then key.cache_key
-        when key.is_a?(Array)                         then key.map { |element| cache_key_with_version(element) }.to_param
-        when key.respond_to?(:to_a)                   then cache_key_with_version(key.to_a)
-        else                                               key.to_param
-        end.to_s
-      end
-
-      def cache_key_with_version(key = @key_parts)
-        self.class.cache_key_with_version(key)
+      def cache_key_with_version
+        CACHE_METHOD_OBJ.send(:retrieve_cache_key, @key_parts)
       end
 
       def update(key)
         @key_parts << key
-        internal_update(key)
+        @cache_key = nil
+        @cache_version = nil
         self
       end
       alias :<< :update
-
-      private
-        def internal_update(key)
-          if key.respond_to?(:cache_key)
-            update_version(key)
-            update_key(key.cache_key.to_s)
-          else
-            split_key(key)
-          end
-        end
-
-        def update_key(key)
-          @cache_key << "/" unless @empty_key
-          @cache_key << key.to_param.to_s
-          @empty_key = false
-        end
-
-        def update_version(key)
-          if key.respond_to?(:cache_version)
-            @cache_version << "/" unless @cache_version.empty?
-            @cache_version << key.cache_version.to_param.to_s
-          else
-            split_version(key)
-          end
-        end
-
-        def split_key(key)
-          case key
-          when Array
-            key.each { |k| internal_update(k) }
-          when Hash
-            key = key.to_a
-            update_version(key)
-            key.sort_by! { |k, _| k.to_s }
-            key.each { |k, v| update_key("#{k}=#{v}") }
-          else
-            if key.respond_to?(:to_a) && !key.nil?
-              internal_update(key.to_a)
-            else
-              update_version(key)
-              update_key(key)
-            end
-          end
-        end
-
-        def split_version(key)
-          if key.is_a?(Array)
-            key.each { |k| update_version(k) }
-          elsif key.respond_to?(:to_a)
-            split_version(key.to_a)
-          end
-        end
     end
   end
 end