This is a "safer" but slower alternative to #36310.

```
Then: 5.97/3.29 => 1.81x faster
Now: 5.97/4.02 => 1.48x faster
```

So the change in performance from #36310 to this PR is  about

```
1 - 3.29/4.02 => 18% # slower than #36310 but still 1.48x faster than master.
```

In #36310 performance was the main goal. In this PR the primary goal was to change as little as possible, performance was the secondary goal. The main difference between the two PRs comes from how `cache_version` and `cache_key` are computed. In #36310 that logic was ported over to the `Cache::Key` class where it could be optimized, in this PR the logic was not moved and instead the `Cache::Key` class leverages the `expanded_version` method in their original location.

How does this PR improve performance for rendering a partial and passing in a collection? 

- On master when a key is used in the cache it has to be generated 3 different times. By using the `Cache::Key` class, it is only generated once and then re-used
- On master, the key that represents the view partial that is being rendered is calculated for every object being pulled from cache inside of the loop, in this PR that view key is generated once outside the loop and re-used.

For safety: In addition to not changing the location of the `expanded_version` and `expanded_key` logic, this class will respect any modified logic of the `Rails.cache` store. So if a cache store customizes their version of `expanded_version` then that version will be used when generating the cache.

Here's some raw numbers from the benchmark provided in #36310 when applied to this PR.

```
$ bundle exec ruby perf_scripts/partial_cache.rb
-- create_table(:customers, {:force=>true})
   -> 0.0115s
/Users/rschneeman/Documents/projects/codetriage/perf_scripts/partial_cache
<#ActiveSupport::Cache::MemoryStore entries=1000, size=352893, options={}>
Warming up --------------------------------------
collection render: no cache
                        20.000  i/100ms
collection render: with cache
                         4.000  i/100ms
Calculating -------------------------------------
collection render: no cache
                        199.729  (± 9.0%) i/s -      1.000k in   5.051812s
collection render: with cache
                         49.642  (± 4.0%) i/s -    248.000  in   5.008595s

Comparison:
collection render: no cache:      199.7 i/s
collection render: with cache:       49.6 i/s - 4.02x  slower
```

> Note that while it's still slower, the prior version was 5.97x slower.