@schneems 2019-12-31
Thanks for the comments I've not been able to re-visit this PR.

As a data point, I did some benchmarking with Derailed against CodeTriage and I was able to make the homepage 1.5x faster by disabling collection cache. I don't think even my patches would have been enough to break even since the partials were so small https://github.com/codetriage/codetriage/pull/1226.

While I do think we can work to make the collection cache code faster, i'm also thinking that the bigger problem people are likely having is that they're using collection caching for tiny partials when they shouldn't.